refactor: Reservation 엔티티에 유효성 검증 로직 추가
